Translate setf+symbol-function assignment to defun
This setf symbol-function was just a experiment in the past to define
functions directly as alias or by defining lambdas, but this is just
too evil. The SBCL linter complains about style-warning of this
operators and functions being not defined because these weird
way to define functions.

To avoid all this mess, I'll just use the canonical way to define
functions with defun on Common Lisp.